requirements.txt
APScheduler==3.10.4
blinker==1.7.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.5.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
nulltype==2.3.1
packaging==23.2
plaid-python==18.4.0
pycurl==7.45.2
pymongo==4.6.1
pyparsing==3.1.1
PySimpleSOAP==1.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
reportbug==0.0.129
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.0
Werkzeug==3.0.1
WTForms==3.1.2
